Move info bar triggers to file open, instead of project open. Fix string compare of package name to be case insensitive.
