+++ Python/Tests/Core/PythonProjectTests.cs
@@ -18,11 +18,9 @@ extern alias pythontools;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using Microsoft.PythonTools.Analysis;
 using Microsoft.PythonTools.Infrastructure;
 using Microsoft.PythonTools.Interpreter;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -41,164 +39,6 @@ namespace PythonToolsTests {
         }
 
         public TestContext TestContext { get; set; }
-
-        [TestMethod, Priority(0)]
-        public void MergeRequirements() {
-            // Comments should be preserved, only package specs should change.
-            AssertUtil.AreEqual(
-                PythonProjectNode.MergeRequirements(new[] {
-                    "a # with a comment",
-                    "B==0.2",
-                    "# just a comment B==01234",
-                    "",
-                    "x < 1",
-                    "d==1.0 e==2.0 f==3.0"
-                }, new[] {
-                    "b==0.1",
-                    "a==0.2",
-                    "c==0.3",
-                    "e==4.0",
-                    "x==0.8"
-                }.Select(p => PackageSpec.FromRequirement(p)), false),
-                "a==0.2 # with a comment",
-                "b==0.1",
-                "# just a comment B==01234",
-                "",
-                "x==0.8",
-                "d==1.0 e==4.0 f==3.0"
-            );
-
-            // addNew is true, so the c==0.3 should be added.
-            AssertUtil.AreEqual(
-                PythonProjectNode.MergeRequirements(new[] {
-                    "a # with a comment",
-                    "b==0.2",
-                    "# just a comment B==01234"
-                }, new[] {
-                    "B==0.1",   // case is updated
-                    "a==0.2",
-                    "c==0.3"
-                }.Select(p => PackageSpec.FromRequirement(p)), true),
-                "a==0.2 # with a comment",
-                "B==0.1",
-                "# just a comment B==01234",
-                "c==0.3"
-            );
-
-            // No existing entries, so the new ones are sorted and returned.
-            AssertUtil.AreEqual(
-                PythonProjectNode.MergeRequirements(null, new[] {
-                    "b==0.2",
-                    "a==0.1",
-                    "c==0.3"
-                }.Select(p => PackageSpec.FromRequirement(p)), false),
-                "a==0.1",
-                "b==0.2",
-                "c==0.3"
-            );
-
-            // Check all the inequalities
-            const string inequalities = "<=|>=|<|>|!=|==";
-            AssertUtil.AreEqual(
-                PythonProjectNode.MergeRequirements(
-                    inequalities.Split('|').Select(s => "a " + s + " 1.2.3"),
-                    new[] { "a==0" }.Select(p => PackageSpec.FromRequirement(p)),
-                    false
-                ),
-                inequalities.Split('|').Select(_ => "a==0").ToArray()
-            );
-        }
-
-        [TestMethod, Priority(0)]
-        public void MergeRequirementsMismatchedCase() {
-            AssertUtil.AreEqual(
-                PythonProjectNode.MergeRequirements(new[] {
-                    "aaaaaa==0.0",
-                    "BbBbBb==0.1",
-                    "CCCCCC==0.2"
-                }, new[] {
-                    "aaaAAA==0.1",
-                    "bbbBBB==0.2",
-                    "cccCCC==0.3"
-                }.Select(p => PackageSpec.FromRequirement(p)), false),
-                "aaaAAA==0.1",
-                "bbbBBB==0.2",
-                "cccCCC==0.3"
-            );
-
-            // https://pytools.codeplex.com/workitem/2465
-            AssertUtil.AreEqual(
-                PythonProjectNode.MergeRequirements(new[] {
-                    "Flask==0.10.1",
-                    "itsdangerous==0.24",
-                    "Jinja2==2.7.3",
-                    "MarkupSafe==0.23",
-                    "Werkzeug==0.9.6"
-                }, new[] {
-                    "flask==0.10.1",
-                    "itsdangerous==0.24",
-                    "jinja2==2.7.3",
-                    "markupsafe==0.23",
-                    "werkzeug==0.9.6"
-                }.Select(p => PackageSpec.FromRequirement(p)), false),
-                "flask==0.10.1",
-                "itsdangerous==0.24",
-                "jinja2==2.7.3",
-                "markupsafe==0.23",
-                "werkzeug==0.9.6"
-            );
-        }
-
-        [TestMethod, Priority(0)]
-        public void FindRequirementsRegexTest() {
-            var r = PythonProjectNode.FindRequirementRegex;
-            AssertUtil.AreEqual(r.Matches("aaaa bbbb cccc").Cast<Match>().Select(m => m.Value),
-                "aaaa",
-                "bbbb",
-                "cccc"
-            );
-            AssertUtil.AreEqual(r.Matches("aaaa#a\r\nbbbb#b\r\ncccc#c\r\n").Cast<Match>().Select(m => m.Value),
-                "aaaa",
-                "bbbb",
-                "cccc"
-            );
-
-            AssertUtil.AreEqual(r.Matches("a==1 b!=2 c<=3").Cast<Match>().Select(m => m.Value),
-                "a==1",
-                "b!=2",
-                "c<=3"
-            );
-
-            AssertUtil.AreEqual(r.Matches("a==1 b!=2 c<=3").Cast<Match>().Select(m => m.Groups["name"].Value),
-                "a",
-                "b",
-                "c"
-            );
-
-            AssertUtil.AreEqual(r.Matches("a==1#a\r\nb!=2#b\r\nc<=3#c\r\n").Cast<Match>().Select(m => m.Value),
-                "a==1",
-                "b!=2",
-                "c<=3"
-            );
-
-            AssertUtil.AreEqual(r.Matches("a == 1 b != 2 c <= 3").Cast<Match>().Select(m => m.Value),
-                "a == 1",
-                "b != 2",
-                "c <= 3"
-            );
-
-            AssertUtil.AreEqual(r.Matches("a == 1 b != 2 c <= 3").Cast<Match>().Select(m => m.Groups["name"].Value),
-                "a",
-                "b",
-                "c"
-            );
-
-            AssertUtil.AreEqual(r.Matches("a -u b -f:x c").Cast<Match>().Select(m => m.Groups["name"].Value),
-                "a",
-                "b",
-                "c"
-            );
-        }
 
         [TestMethod, Priority(0)]
         public void UpdateWorkerRoleServiceDefinitionTest() {
