+++ Python/Tests/VSInterpretersTests/VSInterpretersTests.csproj
@@ -54,6 +54,7 @@
     <Reference Include="Microsoft.VisualStudio.QualityTools.UnitTestFramework" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="PipRequirementsUtilsTests.cs" />
     <Compile Include="StringListReaderTests.cs" />
     <Compile Include="VSInterpretersTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
